refactor: move tests to their own module
